+++ examples/web/bridge/index.ts
@@ -217,72 +217,6 @@ export class BridgeController {
     }
   }
 
-  /**
-   * Connect an Ethereum wallet directly from a private key (dev mode).
-   * Bypasses Reown/WalletConnect — useful for local testing.
-   */
-  async connectEthereumWalletFromKey(
-    privateKey: string,
-    rpcUrl: string,
-    chainId: number
-  ): Promise<void> {
-    try {
-      const { JsonRpcProvider, Wallet } = await import("ethers");
-      const provider = new JsonRpcProvider(rpcUrl, chainId);
-      const signer = new Wallet(privateKey, provider);
-      const address = await signer.getAddress();
-
-      // Create a minimal EIP-1193 provider wrapper for ConnectedEthereumWallet
-      const eip1193Provider = {
-        request: async (args: { method: string; params?: unknown[] }) => {
-          if (args.method === "eth_chainId") {
-            return `0x${chainId.toString(16)}`;
-          }
-          if (
-            args.method === "eth_accounts" ||
-            args.method === "eth_requestAccounts"
-          ) {
-            return [address];
-          }
-          return provider.send(args.method, args.params ?? []);
-        },
-      };
-
-      const wallet = await ConnectedEthereumWallet.from(
-        {
-          chain: ExternalChain.ETHEREUM,
-          provider: eip1193Provider,
-          address,
-          chainId: `0x${chainId.toString(16)}`,
-        },
-        this.chainId
-      );
-
-      // Patch the wallet config to use the private key signer directly
-      // (BrowserProvider from EIP-1193 won't have signing capability for raw keys)
-      const originalToConfig = wallet.toEthWalletConfig.bind(wallet);
-      wallet.toEthWalletConfig = async (ethereumRpcUrl?: string) => {
-        const config = await originalToConfig(ethereumRpcUrl);
-        // Replace the BrowserProvider signer with our direct Wallet signer
-        return { ...config, signer: signer as unknown as typeof config.signer };
-      };
-
-      this.state.connectedEthWallet = wallet;
-      this.revalidateRoute();
-      this.log(
-        `Ethereum wallet connected (dev): ${address.slice(
-          0,
-          6
-        )}...${address.slice(-4)}`,
-        "success"
-      );
-      this.render();
-      this.fetchTokens();
-    } catch (err) {
-      this.log(`Failed to connect Ethereum wallet from key: ${err}`, "error");
-    }
-  }
-
   disconnectEthWallet(): void {
     this.state.connectedEthWallet = undefined;
     this.state.externalBalance = null;
